t: introduce TEST_SCRATCH_DIRECTORY env var to set trash root
Add a TEST_SCRATCH_DIRECTORY environment variable with the same semantics
as the `--root` option to the test harness. This makes it easier to
provide a faster scratch space for the test or the perf suite while
still being able to run tests via `make test` or `make perf` instead of
manually invoking the test driver.